+++ turkdemy/urls.py
@@ -4,6 +4,8 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import URLPattern, URLResolver, include, path
 
+from apps.core.views import switch_language
+
 admin.site.site_header = "TurkDemy Administration"
 admin.site.site_title = "TurkDemy Admin"
 admin.site.index_title = "Operations"
@@ -12,7 +14,7 @@ handler404 = "apps.core.views.page_not_found"
 
 
 urlpatterns: list[URLPattern | URLResolver] = [
-    path("i18n/", include("django.conf.urls.i18n")),
+    path("i18n/setlang/", switch_language, name="set_language"),
     path("health/", include("apps.health.urls")),
     path("api/v1/", include("apps.api.urls")),
     path("accounts/", include("apps.accounts.urls")),